+++ kamehouse-shell/bin/kamehouse/docker/docker-container/docker-sync-ssh-keys-with-host.sh
@@ -15,11 +15,12 @@ fi
 
 mainProcess() {
   ssh-keyscan ${DOCKER_HOST_IP} >> ~/.ssh/known_hosts
+  setIsLinuxDockerHost
+  IS_REMOTE_LINUX_HOST=${IS_LINUX_DOCKER_HOST}
   SSH_USER="${DOCKER_HOST_USERNAME}"
   SSH_SERVER="${DOCKER_HOST_IP}"
   SSH_COMMAND="echo 'ssh from docker container to host successful'"
-  IS_REMOTE_LINUX_HOST=true
-  executeSshCommand "true"
+  executeSshCommand --skip-exit-code-check
 }
 
 main "$@"